+++ MAMBY.Api/Controllers/SaleController.cs
@@ -20,24 +20,50 @@ namespace MAMBY.Api.Controllers
         public async Task<IActionResult> GetAllSale()
         {
             var sales = await _saleService.GetAll();
+            if (sales == null)
+            {
+                return NotFound();
+            }
             return Ok(sales);
         }
         [HttpGet("GetSaleById")]
         public async Task<IActionResult> GetSaleById(int id)
         {
             var sale = await _saleService.GetById(id);
+            if (sale == null)
+            {
+                return NotFound();
+            }
             return Ok(sale);
         }
         [HttpGet("DeleteSale")]
         public IActionResult DeleteSale(int id)
         {
-            _saleService.Delete(id);
-            return Ok();
+            try
+            {
+                if (id == 0)
+                {
+                    return NotFound();
+                }
+                _saleService.Delete(id);
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+
+               return BadRequest(ex.Message);
+            }
+           
+          
         }
         [HttpPost("CreateSale")]
         public async Task<IActionResult> CreateSale(SaleDto saleDto)
         {
             var sale = await _saleService.Create(saleDto);
+            if (sale == null)
+            {
+                return NotFound();
+            }
             return Ok(sale);
         }
     }
